Move vudrm config
